Use cypress run events to signal the test run to VRT
In my opinion it would make sense if the plugin automatically starts and stops when the cypress run starts and stops.

My suggestion would be to use the cypress events to do it. Makes the usage much simpler and also prevents the VRT plugin turn on / off the progress indicator for every spec. 

Whats your opinion on this?